Entities now are obstacles

+++ logic/entities/Entity.py
@@ -35,13 +35,13 @@ class Entity:
     def move(self, radius=1): # TODO: dir as radians
 
         if self.__direction == "left":
-            self.__posPixel.x -= radius*self.__speed
+            self.__posPixel.x -= int(radius*self.__speed)
         elif self.__direction == "right":
-            self.__posPixel.x += radius*self.__speed
+            self.__posPixel.x += int(radius*self.__speed)
         elif self.__direction == "up":
-            self.__posPixel.y -= radius*self.__speed
+            self.__posPixel.y -= int(radius*self.__speed)
         elif self.__direction == "down":
-            self.__posPixel.y += radius*self.__speed
+            self.__posPixel.y += int(radius*self.__speed)
 
     def set_action(self, status):
